Rename shared memory impl. files
